requirements.txt
pymupdf>=1.24.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
